userSignUp connection to backEnd created. _sign-up.scss file moved to /sign-up-form/components. nameAttribute added to SignUp and Login inputElements.
