+++ client/src/features/sign-up-form/components/_sign-up.scss
